Fixed DSP to show correctly
Fixed the DSP to display the visualization correctly for all frequency
bands requested. There are random glitches in decoding or in calculating
the frequency value, which cause the array lookups to overflow
considerably, so the safety checks in the HandleData function need to
sty until we have time to figure out what's going wrong with it.