Changed my mind about how the project should work.
Well, I changed my mind about how the project should work.
Instead of writing each component myself, I took advantage of some of
the Common Controls which windows offers.

I will drop in a toolkit at some point so that this app can be
cross-compiled to linux. But for now this is good enough to use.