Added more Visualizations.
Added Fractal Visualization, which I don't like that much - it takes up
too much CPU.

Added Flame Visualization, which I like quite a lot - it looks more like
a visualization than any of the others because I finally took the time
to rewrite one of the memcpy routines properly.

Still need to add functionality behind the interface, such as looping
and shuffle. Also need to add a way to control the volume and track
position.