Visualizations complete.
Linked visualization loading code to the GUI front end. Loaded
visualizations now show up in the Visualizations menu, along with a
"none" menu item to switch them off for some reason.

Users can now create their own visualizations by copying and modifying
one of the provided plugins, then placing the resulting .dll file
directly inside the "Effects" folder. This requires a restart of the
program in order for the new plugin to be loaded - just be sure that you
do not cause a name conflict, the algorithm will always use the first
item it finds with the given name.